feat(ios): iPad split view, app icon, README
- CalendarView now picks NavigationSplitView on regular width (iPad) and
  keeps NavigationStack on compact (iPhone), sharing all list/filter-bar/
  search/empty-state logic via a new EventListView subview. The split
  view's detail column hosts its own NavigationStack so EventDetailView's
  toolbar renders even with nothing pushed onto it.
- Added a 1024x1024 opaque sRGB app icon (Assets.xcassets/AppIcon.appiconset)
  generated programmatically with CoreGraphics after the SVG-based
  qlmanage render came out letterboxed/transparent.
- Added ios/README.md: how to open/run/debug in Xcode, CLI build/test
  commands, architecture overview + folder map, data endpoints, and
  caching semantics (1h TTL + ETag, 3s non-fatal sidecar timeout, 30-day
  user-state expiry). Notes that CI does not build iOS.
