+++ ios/ChqCalendar/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,1 +1,1 @@
-{"images":[{"idiom":"universal","platform":"ios","size":"1024x1024"}],"info":{"author":"xcode","version":1}}
+{"images":[{"idiom":"universal","platform":"ios","size":"1024x1024","filename":"AppIcon.png"}],"info":{"author":"xcode","version":1}}
