+++ ios/ChqCalendar/Features/Calendar/EventListView.swift
@@ -1,0 +1,175 @@
+import SwiftUI
+
+/// The day-grouped event list shared by both the compact (iPhone,
+/// `NavigationStack`) and regular (iPad, `NavigationSplitView`) layouts in
+/// `CalendarView`.
+///
+/// `selection` is the single knob that changes behavior between the two:
+/// - `nil` (compact): rows are `NavigationLink(value:)`, tapping one pushes
+///   `EventDetailView` onto the enclosing `NavigationStack` via the
+///   `navigationDestination(for:)` below.
+/// - non-nil (regular): rows are plain, `.tag`ged views and the `List`
+///   itself drives selection — tapping one updates the bound `Event?` so the
+///   split view's detail column can show it, no push involved. The
+///   `navigationDestination(for:)` modifier is harmless-but-unused in this
+///   mode since nothing ever pushes a value.
+///
+/// Everything else — the loading/offline/error/no-matches states, the
+/// countdown/offline banners, the filter-bar `safeAreaInset`, and
+/// `refreshable` — is identical between the two layouts, which is the whole
+/// point of sharing this view.
+struct EventListView: View {
+    @Bindable var model: AppModel
+    var selection: Binding<Event?>?
+
+    var body: some View {
+        content
+            .navigationTitle("Chautauqua Calendar")
+            .toolbar { toolbarContent }
+            .safeAreaInset(edge: .top) {
+                // Only shown once there's a snapshot to filter/count
+                // against — during initial launch (no snapshot yet) or the
+                // offline/error empty states, category/location counts
+                // would be meaningless and the bar would just be dead
+                // chrome above a loading spinner or banner.
+                if model.snapshot != nil {
+                    FilterBarView(model: model)
+                }
+            }
+    }
+
+    @ViewBuilder
+    private var content: some View {
+        if model.snapshot == nil {
+            switch model.phase {
+            case .offline:
+                offlineUnavailableView
+            case .failed(let message):
+                errorUnavailableView(message)
+            default:
+                ProgressView("Loading events…")
+            }
+        } else if model.dayGroups.isEmpty {
+            noMatchesView
+        } else {
+            list
+        }
+    }
+
+    private var list: some View {
+        List(selection: selection) {
+            if let days = model.countdownDays {
+                CountdownBanner(days: days)
+            }
+            if model.lastRefreshFailed {
+                OfflineBanner()
+            }
+
+            ForEach(model.dayGroups) { day in
+                Section {
+                    ForEach(day.events) { event in
+                        row(for: event)
+                    }
+                } header: {
+                    dayHeader(for: day)
+                }
+            }
+
+            if model.filter.dateScope == .next {
+                Button("Show next day") {
+                    model.showNextDay()
+                }
+            }
+        }
+        .listStyle(.plain)
+        .refreshable {
+            await model.refresh(force: true)
+        }
+        .navigationDestination(for: Event.self) { event in
+            EventDetailView(event: event, model: model)
+        }
+    }
+
+    @ViewBuilder
+    private func row(for event: Event) -> some View {
+        if selection != nil {
+            EventRow(model: model, event: event)
+                .tag(event)
+        } else {
+            NavigationLink(value: event) {
+                EventRow(model: model, event: event)
+            }
+        }
+    }
+
+    private func dayHeader(for day: DayGroup) -> some View {
+        HStack {
+            Text(day.title)
+            Spacer()
+            ForEach(day.weekNumbers, id: \.self) { number in
+                Text("Wk \(number)")
+                    .font(.caption2.weight(.semibold))
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 2)
+                    .background(.secondary.opacity(0.15), in: Capsule())
+            }
+        }
+    }
+
+    private var noMatchesView: some View {
+        ContentUnavailableView {
+            Label("No matching events", systemImage: "calendar.badge.exclamationmark")
+        } actions: {
+            Button("Clear Filters") {
+                model.clearFilters()
+            }
+        }
+    }
+
+    private var offlineUnavailableView: some View {
+        ContentUnavailableView {
+            Label("You're Offline", systemImage: "wifi.slash")
+        } description: {
+            Text("Connect to the internet to load this season's events.")
+        } actions: {
+            Button("Retry") {
+                Task { await model.refresh(force: true) }
+            }
+            .disabled(model.isRefreshing)
+        }
+    }
+
+    private func errorUnavailableView(_ message: String) -> some View {
+        ContentUnavailableView {
+            Label("Something Went Wrong", systemImage: "exclamationmark.triangle")
+        } description: {
+            Text(message)
+        } actions: {
+            Button("Retry") {
+                Task { await model.refresh(force: true) }
+            }
+            .disabled(model.isRefreshing)
+        }
+    }
+
+    @ToolbarContentBuilder
+    private var toolbarContent: some ToolbarContent {
+        ToolbarItem(placement: .topBarTrailing) {
+            Menu {
+                ForEach(model.years, id: \.self) { year in
+                    Button {
+                        Task { await model.select(year: year) }
+                    } label: {
+                        if year == model.selectedYear {
+                            Label(String(year), systemImage: "checkmark")
+                        } else {
+                            Text(String(year))
+                        }
+                    }
+                }
+            } label: {
+                Text(String(model.selectedYear))
+            }
+        }
+    }
+}
